--- components/menu-site/Footer.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import 'bootstrap/dist/css/bootstrap.css';
-
-function Footer(props) {
-    return (
-        <div className="container-fluid p-3"  style={{background:" #4682B4"}}>
-            <p className="text-center text-light">SGME-2024</p>
-        </div>
-    );
-}
-
-export default Footer;
