+++ package.json
@@ -9,9 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@heroicons/react": "^2.1.3",
+    "@next/font": "^14.2.3",
     "axios": "^1.6.0",
     "bootstrap": "^5.3.2",
-    "bootstrap-icons": "^1.11.1",
+    "bootstrap-icons": "^1.11.3",
     "date-fns": "^3.0.6",
     "framer-motion": "^10.16.4",
     "js-cookie": "^3.0.5",
@@ -21,6 +23,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.48.2",
+    "react-icons": "^5.2.1",
     "react-imask": "^7.3.0",
     "react-input-mask": "^2.0.4",
     "react-modal": "^3.16.1",
