+++ styles/globals.css
@@ -1,10 +1,16 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap');
+
 *{
     margin: 0;
     padding: 0;
     box-sizing: border-box;
     color: #393939;
-    font-size: 16px;
+    font-family: 'Roboto', sans-serif;
+    font-weight:400 ;
+    font-size: 17px;
+
 }
+
 
 .style-heigth-90vh{
     height: 90vh;
